+++ src/main/java/com/hello/boilerplate/domain/auth/controller/AuthController.java
@@ -6,8 +6,8 @@ import com.hello.boilerplate.domain.auth.dto.ReissueResponseDto;
 import com.hello.boilerplate.domain.auth.service.AuthService;
 import com.hello.boilerplate.domain.auth.utils.JwtUtil;
 import com.hello.boilerplate.domain.user.domain.User;
-import com.hello.boilerplate.global.dto.ApiResponse;
-import com.hello.boilerplate.global.infrastructure.CookieUtil;
+import com.hello.boilerplate.global.presentation.dto.ApiResponse;
+import com.hello.boilerplate.global.infrastructure.web.CookieUtil;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
