+++ src/main/java/com/hello/boilerplate/domain/auth/service/AuthService.java
@@ -7,8 +7,8 @@ import com.hello.boilerplate.domain.auth.exception.AuthorizationErrorMessages;
 import com.hello.boilerplate.domain.auth.utils.JwtUtil;
 import com.hello.boilerplate.domain.user.domain.User;
 import com.hello.boilerplate.domain.user.domain.UserRepository;
-import com.hello.boilerplate.global.exception.CustomException;
-import com.hello.boilerplate.global.exception.UnauthorizedException;
+import com.hello.boilerplate.global.common.exception.CustomException;
+import com.hello.boilerplate.global.common.exception.UnauthorizedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
