+++ src/main/java/com/hello/boilerplate/domain/auth/utils/JwtUtil.java
@@ -4,7 +4,7 @@ import com.hello.boilerplate.domain.auth.constants.JwtConstants;
 import com.hello.boilerplate.domain.auth.exception.AuthorizationErrorMessages;
 import com.hello.boilerplate.domain.auth.service.RedisTokenService;
 import com.hello.boilerplate.domain.user.domain.User;
-import com.hello.boilerplate.global.exception.UnauthorizedException;
+import com.hello.boilerplate.global.common.exception.UnauthorizedException;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
