+++ src/main/java/com/hello/boilerplate/global/common/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.hello.boilerplate.global.exception;
+package com.hello.boilerplate.global.common.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
